bcrypt beginnings
