+++ server/package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
     "dot-env": "^0.0.1",
     "dotenv": "^16.4.5",
